Initial commit
This is the wikibase-registry setup mostly as I found it on the server,
set up and maintained by others so far. I moved the database settings
from LocalSettings.php to PrivateSettings.php, and from
docker-compose.yml to .env. I also added a LICENSE in accordance with
our GitHub guidelines.

I am not sure if the entrypoint.sh is actually used, but it’s here, so
let’s commit it. We can always remove it later.